started windows #foreign support -- floating point not working yet--also found bug with eval EXPR_CALL

+++ infer.c
@@ -18,7 +18,7 @@ static bool infer_from_expr(Typer *tr, Expression *match, Value to, Type *to_typ
 		/* an identifier! maybe it's one of idents... */
 		arr_foreach(idents, Identifier, ident) {
 			if (ident_eq_string(*ident, match->ident_str)) {
-				long idx = ident - idents;
+				long idx = (long)(ident - idents);
 				types[idx] = *to_type;
 				vals[idx] = to;
 				break;
